pyproject.toml
[project]
name = "pg2-benchmark"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "tintinrevient", email = "zhaobenben007@googlemail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "typer>=0.15.2",
    "pg2-dataset",
    "python-on-whales>=0.76.1",
    "gitpython>=3.1.44",
    "pycm>=4.3",
    "scipy>=1.15.3",
    "dvc>=3.59.2",
]

[project.scripts]
pg2-benchmark = "pg2_benchmark.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
pg2-dataset = { git = "https://github.com/ProteinGym2/pg2-dataset.git", branch = "main" }

[tool.ruff]
line-length = 88

[tool.click]
logging_level = "INFO"
logging_format = "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
